@@ -32,9 +32,9 @@
     "browserslist": "^4.20.2",
     "file-saver": "^2.0.5",
     "jszip": "^3.7.1",
-    "react": "^17.0.2",
+    "react": "^18.0.0",
     "react-bootstrap": "^2.2.2",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.0.0",
     "react-router-dom": "^6.2.2",
     "react-select": "^5.2.2",
     "react-string-replace": "^1.0.0",
